requirements.txt
streamlit==1.37.1
streamlit-webrtc==0.47.7
opencv-python==4.10.0.84
numpy==1.26.4
av==12.3.0
mediapipe==0.10.14
pillow==10.4.0